ray sphere intersection

+++ src/maths/vec3.h
@@ -6,22 +6,22 @@ using std::sqrt;
 class vec3{
 
     public:
-        float e[3];
+        double e[3];
 
     public:
         vec3();
-        vec3(float e0, float e1, float e2);
-        float x() const {return e[0];}
-        float y() const {return e[1];}
-        float z() const {return e[2];}
+        vec3(double e0, double e1, double e2);
+        double x() const {return e[0];}
+        double y() const {return e[1];}
+        double z() const {return e[2];}
 
         vec3 operator-() const;
-        float& operator[](int i);
+        double& operator[](int i);
         vec3& operator+=(const vec3 &v);
-        vec3& operator*=(const float t);
-        vec3& operator/=(const float t);
-        float length() const;
-        float length_squared() const;
+        vec3& operator*=(const double t);
+        vec3& operator/=(const double t);
+        double length() const;
+        double length_squared() const;
 
 
 
@@ -50,19 +50,19 @@ inline vec3 operator*(const vec3 &u, const vec3 &v) {
     return vec3(u.e[0] * v.e[0], u.e[1] * v.e[1], u.e[2] * v.e[2]);
 }
 
-inline vec3 operator*(float t, const vec3 &v) {
+inline vec3 operator*(double t, const vec3 &v) {
     return vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 }
 
-inline vec3 operator*(const vec3 &v, float t) {
+inline vec3 operator*(const vec3 &v, double t) {
     return t * v;
 }
 
-inline vec3 operator/(vec3 v, float t) {
+inline vec3 operator/(vec3 v, double t) {
     return (1/t) * v;
 }
 
-inline float dot(const vec3 &u, const vec3 &v) {
+inline double dot(const vec3 &u, const vec3 &v) {
     return u.e[0] * v.e[0]
          + u.e[1] * v.e[1]
          + u.e[2] * v.e[2];
